% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_column_types.R
\name{change_column_types}
\alias{change_column_types}
\title{Change column types for social networks and fertility data (LISS: wj18a_EN_1.0p.sav)}
\usage{
change_column_types(data)
}
\arguments{
\item{data}{Tibble which is the result of \code{translate(read_data())}}
}
\value{
Tibble with corrected column types and updated labels
}
\description{
Change column types for social networks and fertility data (LISS: wj18a_EN_1.0p.sav)
}
\examples{
\dontshow{if (file.exists("wj18a_EN_1.0p.sav")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
read_data() |> translate() |> change_column_types()
\dontshow{\}) # examplesIf}
}
