% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alter_attr.R
\name{create_alter_attr}
\alias{create_alter_attr}
\title{Create dataframe of alter attributes based on social networks and fertility data (LISS: wj18a_EN_1.0p.sav)}
\usage{
create_alter_attr(data)
}
\arguments{
\item{data}{Tibble which is the result of \code{create_relation_labels(fix_errors(change_column_types(translate(read_data()))))}}
}
\value{
Tibble with variable \code{alter_attr} which includes a dataframe with alter attributes for each respondent
}
\description{
Create dataframe of alter attributes based on social networks and fertility data (LISS: wj18a_EN_1.0p.sav)
}
\examples{
\dontshow{if (file.exists("wj18a_EN_1.0p.sav")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data <- read_data() |> translate() |>
change_column_types() |> fix_errors() |> create_relation_labels()
create_alter_attr(data[1, ])
\dontshow{\}) # examplesIf}
}
