% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nw.R
\name{create_nw}
\alias{create_nw}
\title{Create dataframes of alter attributes and edgelists and store them in list-columns for social networks and fertility data (LISS: wj18a_EN_1.0p.sav)}
\usage{
create_nw(data)
}
\arguments{
\item{data}{Tibble which is the result of \code{create_relation_labels(fix_errors(change_column_types(translate(read_data()))))}}
}
\value{
Tibble with list-columns containing alter attributes and edgelists
}
\description{
Create dataframes of alter attributes and edgelists and store them in list-columns for social networks and fertility data (LISS: wj18a_EN_1.0p.sav)
}
\examples{
\dontshow{if (file.exists("wj18a_EN_1.0p.sav")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
read_data() |> translate() |> change_column_types() |>
fix_errors() |> create_relation_labels() |> create_nw()
\dontshow{\}) # examplesIf}
}
