% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_relation_labels.R
\name{create_relation_labels}
\alias{create_relation_labels}
\title{Produces corrected relationship labels for social networks and fertility data (LISS: wj18a_EN_1.0p.sav)}
\usage{
create_relation_labels(data)
}
\arguments{
\item{data}{Tibble which is the result of \code{fix_errors(change_column_types(translate(read_data())))}}
}
\value{
Tibble in which data on relationship labels are corrected and improved
}
\description{
Produces corrected relationship labels for social networks and fertility data (LISS: wj18a_EN_1.0p.sav)
}
\examples{
\dontshow{if (file.exists("wj18a_EN_1.0p.sav")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
read_data() |> translate() |> change_column_types() |>
fix_errors() |> create_relation_labels()
\dontshow{\}) # examplesIf}
}
