% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tidygraph.R
\name{create_tidygraph}
\alias{create_tidygraph}
\title{Create tidygraph objects from social networks and fertility data (LISS: wj18a_EN_1.0p.sav)}
\usage{
create_tidygraph(data)
}
\arguments{
\item{data}{Tibble which is the result of \code{produce_data()}}
}
\value{
Tibble with variable \code{tidygraph} that includes tidygraph objects for all respondents
}
\description{
Create tidygraph objects from social networks and fertility data (LISS: wj18a_EN_1.0p.sav)
}
\examples{
\dontshow{if (file.exists("wj18a_EN_1.0p.sav")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
produce_data() |> create_tidygraph()
\dontshow{\}) # examplesIf}
}
