% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_errors.R
\name{fix_errors}
\alias{fix_errors}
\title{Fixes reporting errors and inconsistencies in social networks and fertility data (LISS: wj18a_EN_1.0p.sav)}
\usage{
fix_errors(data)
}
\arguments{
\item{data}{Tibble which is the result of \verb{change_column_types(translate(read_data())}}
}
\value{
Tibble in which data errors are fixed and data worries are flagged
}
\description{
Fixes reporting errors and inconsistencies in social networks and fertility data (LISS: wj18a_EN_1.0p.sav)
}
\examples{
\dontshow{if (file.exists("wj18a_EN_1.0p.sav")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
read_data() |> translate() |> change_column_types() |> fix_errors()
\dontshow{\}) # examplesIf}
}
