% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_background_vars.R
\name{get_background_vars}
\alias{get_background_vars}
\title{Get respondent background variables (LISS: avars_201802_EN_1.0p.sav) for social networks and fertility data (LISS: wj18a_EN_1.0p.sav)}
\usage{
get_background_vars(file = "avars_201802_EN_1.0p.sav")
}
\arguments{
\item{file}{Path to file avars_201802_EN_1.0p.sav (or renamed variant)}
}
\value{
Tibble of data with background variables for social networks and fertility data
}
\description{
Get respondent background variables (LISS: avars_201802_EN_1.0p.sav) for social networks and fertility data (LISS: wj18a_EN_1.0p.sav)
}
\examples{
\dontshow{if (file.exists("avars_201802_EN_1.0p.sav")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_background_vars()
\dontshow{\}) # examplesIf}
}
