% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Reads-in social networks and fertility data (LISS: wj18a_EN_1.0p.sav)}
\usage{
read_data(file = "wj18a_EN_1.0p.sav")
}
\arguments{
\item{file}{Path to file wj18a_EN_1.0p.sav (or renamed variant)}
}
\value{
Tibble of social networks and fertility data
}
\description{
Reads-in social networks and fertility data (LISS: wj18a_EN_1.0p.sav)
}
\examples{
\dontshow{if (file.exists("wj18a_EN_1.0p.sav")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
read_data("wj18a_EN_1.0p.sav")
\dontshow{\}) # examplesIf}
}
