% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_timing_vars.R
\name{remove_timing_vars}
\alias{remove_timing_vars}
\title{Remove variables related to timing of giving answers in survey}
\usage{
remove_timing_vars(data)
}
\arguments{
\item{data}{Tibble sent within function \code{produce_data()}}
}
\value{
Tibble without timing variables
}
\description{
Remove variables related to timing of giving answers in survey
}
