% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate LISS variable names of social networks and fertility data (LISS: wj18a_EN_1.0p.sav) into sensible English names}
\usage{
translate(data)
}
\arguments{
\item{data}{Tibble which is the result of \code{read_data()}}
}
\value{
Tibble with sensible column names
}
\description{
Translate LISS variable names of social networks and fertility data (LISS: wj18a_EN_1.0p.sav) into sensible English names
}
\examples{
\dontshow{if (file.exists("wj18a_EN_1.0p.sav")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
read_data() |> translate()
\dontshow{\}) # examplesIf}
}
